+++ src/controllers/GameHistory.js
@@ -1,0 +1,5 @@
+const GameHistory = require("../db/models/GameHistory");
+
+exports.find = async(search, projection, options = {}) => await GameHistory.find(search, projection, options);
+
+exports.findOne = async(search, projection, options = {}) => await GameHistory.findOne(search, projection, options);