+++ src/helpers/constants/Game.js
@@ -12,15 +12,14 @@ exports.populate = [
     { path: "gameMaster", select: "-password" },
 ];
 
-exports.turnPreActionsOrder = [
+exports.turnPreNighttActionsOrder = [
     { source: "all", action: "elect-mayor" },
 ];
 
-exports.turnActionsOrder = [
+exports.turnNightActionsOrder = [
     { source: "seer", action: "look" },
     { source: "wolves", action: "eat" },
     { source: "witch", action: "use-potion" },
     { source: "protector", action: "protect" },
     { source: "raven", action: "mark" },
-    { source: "all", action: "vote" },
 ];